% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcrlscR.R
\docType{data}
\name{man_qc}
\alias{man_qc}
\title{man_qc: test data for QC-RLSC}
\format{
A list with data matrix and meta data:
\describe{
\item{data}{A data frame with 462 replicates (row) and 656 features
(column)}
\item{meta}{A data frame with 2 columns:
\itemize{
\item batch: 4 batches
\item sample_type: QC and Sample
}
}
}
}
\usage{
man_qc
}
\description{
This HPLC data set includes 4 batches with missing values.
}
\examples{
man_qc
t(sapply(man_qc, dim))
## Select data matrix and meta data
data <- man_qc$data
meta <- man_qc$meta
## Select batches and data types
cls.qc <- factor(meta$sample_type)
cls.bl <- factor(meta$batch)
}
\keyword{datasets}
