% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcrlscR.R
\name{mv.perc}
\alias{mv.perc}
\title{Missing value percentage}
\usage{
mv.perc(x)
}
\arguments{
\item{x}{an vector, matrix or data frame.}
}
\value{
missing value percentage.
}
\description{
Calculate missing value percentage.
}
\examples{
names(man_qc)
data <- man_qc$data
meta <- man_qc$meta
## check missing value rates
tail(sort(mv.perc(data)), 20)
}
\seealso{
Other missing value processing: 
\code{\link{mv.filter}()},
\code{\link{mv.filter.qc}()}
}
\concept{missing value processing}
