% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcrlscR.R
\name{mv_perc}
\alias{mv_perc}
\title{Missing value percentage}
\usage{
mv_perc(x)
}
\arguments{
\item{x}{an vector, matrix or data frame.}
}
\value{
missing value percentage.
}
\description{
Calculate missing value percentage.
}
