% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcrlscR.R
\name{qc_rlsc}
\alias{qc_rlsc}
\title{QC based robust LOESS signal correction (QC-RLSC)}
\usage{
qc_rlsc(x, y, method = c("subtract", "divide"), opti = TRUE, ...)
}
\arguments{
\item{x}{A data frame with samples (row) and variables (column).}

\item{y}{A vector with string of "qc" and "sample".}

\item{method}{Data scaling method.}

\item{opti}{A logical value indicating whether or not optimise 'span'}

\item{...}{Other parameter for 'loess'.}
}
\value{
A corrected data frame.
}
\description{
QC based robust LOESS (locally estimated scatterplot smoothing) signal
correction (QC-RLSC)
}
\examples{
names(man_qc)
data <- man_qc$data
meta <- man_qc$meta

cls.qc <- factor(meta$sample_type)
cls.bl <- factor(meta$batch)

## apply QC-RLSC with optimisation of 'span'
res_1 <- qc_rlsc(data, cls.qc, method = "subtract", opti = TRUE)

\dontrun{
## Use PCA and PCA-LDA to check. Use R package 'mt' here. 

## install.packages("mt")
library("mt")

res_fill <- res_1

res_fill <- mv.fill(res_fill, method = "median", ze_ne = TRUE)
pcaplot(res_fill, cls.qc, pcs = c(2, 1), ep = 1)
pcaplot(res_fill, cls.bl, pcs = c(2, 1), ep = 1)
plot(pcalda(res_fill, cls.bl))
}

## apply QC-RLSC without optimisation of 'span'
res_2 <- qc_rlsc(data, cls.qc, method = "subtract", opti = FALSE)

\dontrun{
res_fill <- res_2
res_fill <- mv.fill(res_fill, method = "median", ze_ne = TRUE)
pcaplot(res_fill, cls.qc, pcs = c(2, 1), ep = 1)
pcaplot(res_fill, cls.bl, pcs = c(2, 1), ep = 1)
plot(pcalda(res_fill, cls.bl))
}
}
\references{
Dunn et al. Procedures for large-scale metabolic profiling of serum and
plasma using gas chromatography and liquid chromatography coupled to
mass spectrometry. Nature Protocols 6, 1060–1083 (2011)
}
