% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcrlscR.R
\name{qc_rlsc_wrap}
\alias{qc_rlsc_wrap}
\title{Wrapper function for QC-RLSC}
\usage{
qc_rlsc_wrap(
  dat,
  cls.qc,
  cls.bl,
  method = c("subtract", "divide"),
  intra = FALSE,
  opti = TRUE,
  log10 = TRUE,
  outl = TRUE,
  shift = TRUE,
  ...
)
}
\arguments{
\item{dat}{A data frame with samples (row) and variables (column).}

\item{cls.qc}{A vector with string of "qc" and "sample".}

\item{cls.bl}{A vector with string of batch indicators.}

\item{method}{Data scaling method. Support "subtract" and "divide"}

\item{intra}{A logical value indicating whether signal correction is
performed inside each batch.}

\item{opti}{A logical value indicating whether or not 'span' parameters
are optimised.}

\item{log10}{A logical value indicating whether log10 transformation for
the data set or not.}

\item{outl}{A logical value indicating whether or not QC outlier
detection is employed.}

\item{shift}{A logical value indicating whether or not batch shift is
applied after signal correction.}

\item{...}{Other parameter for 'loess'.}
}
\value{
A corrected data frame.
}
\description{
Wrapper function for QC-RLSC
}
