% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcrlscR.R
\docType{package}
\name{qcrlscR-package}
\alias{qcrlscR}
\alias{qcrlscR-package}
\title{qcrlscR: Quality Control–based Robust LOESS Signal Correction in R}
\description{
\code{qcrlscR} implements quality control–based robust LOESS
signal correction for metabolomics data analysis.
}
\section{Main functions}{

The \code{qcrlscR} provides functions for signal correction for metabolomics
data analysis. It allows users to optimise \code{span} for LOESS. This package
also provides simple functions for univariate outlier detection and
missing value filtering. Users need to use other R packages for missing
value filling before applying this package if the data set has missing
values.
}

\section{Package context}{

This package does not use "tidyverse" and only uses basic R for
simplicity and easy maintenance.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wanchanglin/qcrlscR}
  \item Report bugs at \url{https://github.com/wanchanglin/qcrlscR/issues}
}

}
\keyword{internal}
