% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcrlscR.R
\name{mv_perc}
\alias{mv_perc}
\title{Missing value percentage}
\usage{
mv_perc(x)
}
\arguments{
\item{x}{an vector, matrix or data frame.}
}
\value{
missing value percentage.
}
\description{
Calculate missing value percentage.
}
\examples{
names(man_qc)
data <- man_qc$data
meta <- man_qc$meta
## check missing value rates
tail(sort(mv_perc(data)), 20)
}
\seealso{
Other missing value processing: 
\code{\link{mv_filter}()},
\code{\link{mv_filter_qc}()}
}
\concept{missing value processing}
