% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcrlscR.R
\docType{package}
\name{qcrlscR-package}
\alias{qcrlscR}
\alias{qcrlscR-package}
\title{qcrlscR: Quality control–based robust LOESS signal correction in R}
\description{
\code{qcrlscR} implements quality control–based robust LOESS
signal correction for metabolomics data analysis.
}
\section{Main functions}{

The main qcrlscR provides more functions for metabolomics data analysis.
These functions are statistical analysis and plot methods with \code{ggplot2}
and \code{lattice}. It uses \code{tidyverse} as well as \code{reshape2} and \code{plyr}
packages.
}

\section{Package context}{

This package follows the principles of the "tidyverse" as mush as possible.
It also uses \code{melt} in \code{reshape2} if \code{tidyr} is complicated in some
circumstances.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wanchanglin/qcrlscR}
  \item Report bugs at \url{https://github.com/wanchanglin/qcrlscR/issues}
}

}
\keyword{internal}
